heatsim/requirements.txt
# Subseatec Materials Simulation Platform
# Python 3.11+

# Web Framework
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.0
Flask-WTF>=1.2.0
Flask-Migrate>=4.0.0

# Database
SQLAlchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# Security
Werkzeug>=3.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0

# Production Server
gunicorn>=22.0.0

# Data Processing (Phase 2+)
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
openpyxl>=3.1.0

# Visualization (Phase 3+)
matplotlib>=3.7.0

# COMSOL Integration (Phase 4)
mph>=1.2.0                 # Python COMSOL API
pyvista>=0.42.0            # 3D visualization and VTK handling
imageio>=2.31.0            # Image/video I/O
imageio-ffmpeg>=0.4.9      # FFmpeg backend for animations
vtk>=9.2.0                 # VTK for 3D processing
Pillow>=10.0.0             # Image processing

# CAD Geometry
cadquery>=2.4.0             # STEP file analysis

# Report Generation
python-docx>=1.1.0
fpdf2>=2.7.0

# Development
pytest>=7.3.0
pytest-flask>=1.3.0
